Bug 989706: Quiet extra output from Libcgroup.usage
